Move annotations lookup to `io.spine.reflect`
