+++ logging/src/jvmMain/kotlin/io/spine/reflect/PackageAnnotationLookup.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.logging
+package io.spine.reflect
 
 import java.lang.annotation.ElementType
 import java.lang.annotation.Repeatable
